backend/requirements.txt
# =============================================
# HKTVmall AI 營運系統 - Python 依賴
# =============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.31.0
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# External APIs
firecrawl-py==0.0.16
anthropic==0.18.1
boto3==1.34.34

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
croniter==2.0.1

# Data Export
pandas==2.2.0
openpyxl==3.1.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
black==24.1.1
ruff==0.2.0
